+++ src/main/scala/io/archivesunleashed/util/TweetUtils.scala
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.archivesunleashed.matchbox
+package io.archivesunleashed.util
 
 import org.json4s.JsonAST._
