+++ src/test/scala/io/archivesunleashed/matchbox/ExtractEntitiesTest.scala
@@ -21,6 +21,7 @@ import java.io.File
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.module.scala.DefaultScalaModule
 import com.google.common.io.{Files, Resources}
+import io.archivesunleashed.app.ExtractEntities
 import org.apache.commons.io.FileUtils
 import org.apache.commons.logging.LogFactory
 import org.apache.spark.{SparkConf, SparkContext}
