+++ src/test/scala/io/archivesunleashed/matchbox/WriteGEXFTest.scala
@@ -17,12 +17,15 @@
 package io.archivesunleashed.matchbox
 
 
-import org.apache.spark.{ SparkConf, SparkContext }
+import org.apache.spark.{SparkConf, SparkContext}
 import org.junit.runner.RunWith
 import org.scalatest.junit.JUnitRunner
-import org.scalatest.{ BeforeAndAfter, FunSuite }
+import org.scalatest.{BeforeAndAfter, FunSuite}
 import java.io.File
-import java.nio.file.{Paths, Files}
+import java.nio.file.{Files, Paths}
+
+import io.archivesunleashed.app.WriteGEXF
+
 import scala.io.Source
 
 @RunWith(classOf[JUnitRunner])
